requirements.txt
Django>=4.2,<5.0
djangorestframework
django-environ
mongoengine
celery
redis
